requirements.txt
aiogram==3.8.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyYAML>=6.0
jsonschema>=4.0.0
requests>=2.31.0
rich>=13.0
python-dotenv>=1.0
email-validator>=2.1.0